null pointer and invalid command exception

+++ src/main/java/Controller/BossProcessor.java
@@ -2,7 +2,9 @@ package Controller;
 
 import View.BossView;
 import model.*;
+import model.request.Request;
 
+import java.lang.NullPointerException;
 import java.util.ArrayList;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -18,15 +20,7 @@ public class BossProcessor extends Processor {
         return bossViewManager;
     }
 
-    public void viewPersonalInfo(String userName) {
-        //TODO : error handling
-        User user = User.getUserByUserName(userName);
-        bossViewManager.viewUser(user);
-
-    }
-
-    public void editField(String userName, String field, String changeField) {
-        //TODO : error handling
+    public void editField(String userName, String field, String changeField) throws InvalidCommandException {
         User user = User.getUserByUserName(userName);
         field = field.trim();
         Pattern firstNamePattern = Pattern.compile("^(?i)first name$");
@@ -49,11 +43,12 @@ public class BossProcessor extends Processor {
             user.getUserPersonalInfo().setPhoneNumber(changeField);
         } else if (passwordMatcher.matches()) {
             user.getUserPersonalInfo().setPassword(changeField);
+        } else {
+            throw new InvalidCommandException("invalid command");
         }
     }
 
-    public void manageUsers(String command) {
-        //TODO : error handling
+    public void manageUsers(String command) throws InvalidCommandException {
         if (command.equals("back")) {
             return;
         }
@@ -64,25 +59,44 @@ public class BossProcessor extends Processor {
         Pattern createManagerPattern = Pattern.compile("create manager profile");
         Matcher createManagerMatcher = createManagerPattern.matcher(command);
         if (viewUserMatcher.matches()) {
-            processViewUser(viewUserMatcher.group(1));
+            try {
+                processViewUser(viewUserMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (deleteUserMatcher.matches()) {
-            processDeleteUser(deleteUserMatcher.group(1));
+            try {
+                processDeleteUser(deleteUserMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (createManagerMatcher.matches()) {
-            processCreateManagerProfile();
+            try {
+                processCreateManagerProfile();
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else {
+            throw new InvalidCommandException("invalid command");
         }
 
     }
 
-    public void processViewUser(String userName) {
-        //TODO : error handling
+    public void processViewUser(String userName) throws NullPointerException {
         User user = User.getUserByUserName(userName);
+        if (user == null) {
+            throw new NullPointerException("user with this user name doesn't exist");
+        }
+
         bossViewManager.viewUser(user);
 
     }
 
-    public void processDeleteUser(String userName) {
-        //TODO : error handling
+    public void processDeleteUser(String userName) throws NullPointerException {
         User userToBeDeleted = User.getUserByUserName(userName);
+        if (userToBeDeleted == null) {
+            throw new NullPointerException("user with this user name doesn't exist");
+        }
         ((Manager) user).deleteUser(userToBeDeleted);
     }
 
@@ -93,21 +107,28 @@ public class BossProcessor extends Processor {
         ((Manager) user).createManagerProfile(managerInfo);
     }
 
-    public void manageAllProducts(String command) {
-        //TODO : error handling
+    public void manageAllProducts(String command) throws InvalidCommandException {
         if (command.equals("back")) {
             return;
         }
         Pattern removeProductPattern = Pattern.compile("^(?i)remove (.+)$");
         Matcher removeProductMatcher = removeProductPattern.matcher(command);
         if (removeProductMatcher.matches()) {
-            processRemoveProduct(removeProductMatcher.group(1));
+            try {
+                processRemoveProduct(removeProductMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else {
+            throw new InvalidCommandException("invalid command");
         }
     }
 
-    public void processRemoveProduct(String productId) {
-        //TODO : error handling
+    public void processRemoveProduct(String productId) throws NullPointerException {
         Product product = Product.getProductById(productId);
+        if (product == null) {
+            throw new NullPointerException("product with this productId doesn't exist");
+        }
         ((Manager) user).removeProduct(product);
     }
 
@@ -119,8 +140,7 @@ public class BossProcessor extends Processor {
 
     }
 
-    public void manageCodedDiscounts(String command) {
-        //TODO : error handling
+    public void manageCodedDiscounts(String command) throws InvalidCommandException {
         if (command.equals("back")) {
             return;
         }
@@ -131,32 +151,51 @@ public class BossProcessor extends Processor {
         Pattern removeCodedDiscountPattern = Pattern.compile("remove discount code (.+)");
         Matcher removeCodedDiscountMatcher = removeCodedDiscountPattern.matcher(command);
         if (viewCodedDiscountMatcher.matches()) {
-            processViewCodedDiscount(viewCodedDiscountMatcher.group(1));
+            try {
+                processViewCodedDiscount(viewCodedDiscountMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (editCodedDiscountMatcher.matches()) {
-            processEditCodedDiscountFirst(editCodedDiscountMatcher.group(1));
+            try {
+                processEditCodedDiscountFirst(editCodedDiscountMatcher.group(1));
+            } catch (InvalidCommandException | NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (removeCodedDiscountMatcher.matches()) {
-            processRemoveCodedDiscount(removeCodedDiscountMatcher.group(1));
+            try {
+                processRemoveCodedDiscount(removeCodedDiscountMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else {
+            throw new InvalidCommandException("invalid command");
         }
     }
 
-    public void processViewCodedDiscount(String discountCode) {
-        //TODO : error handling
+    public void processViewCodedDiscount(String discountCode) throws NullPointerException {
         CodedDiscount codedDiscount = CodedDiscount.getDiscountById(discountCode);
+        if (codedDiscount == null) {
+            throw new NullPointerException("coded discount with this code doesn't exist");
+        }
         bossViewManager.viewCodedDiscount(codedDiscount);
 
     }
 
-    public void processEditCodedDiscountFirst(String discountCode) {
-        //TODO : error handling
-        ArrayList<String> discountInfo = new ArrayList<>();
-        bossViewManager.getEditCodedDiscountInfo(discountInfo);
-        for (int i = 0; i < discountInfo.size(); i += 2) {
-            processEditCodedDiscountSecond(discountInfo.get(i), discountInfo.get(i + 1), discountCode);
+    public void processEditCodedDiscountFirst(String discountCode) throws NullPointerException, InvalidCommandException {
+        if (CodedDiscount.getDiscountById(discountCode) == null) {
+            throw new NullPointerException("coded discount with this code doesn't exist");
+        }
+        String field = bossViewManager.getEditCodedDiscountField();
+        if (field.equalsIgnoreCase("back")) {
+            return;
+        } else {
+            processEditCodedDiscountSecond(field, discountCode);
         }
     }
 
-    public void processEditCodedDiscountSecond(String field, String changeField, String discountCode) {
-        //TODO : error handling
+    public void processEditCodedDiscountSecond(String field, String discountCode) throws InvalidCommandException {
+        String changeField = bossViewManager.getEditCodedDiscountInField();
         CodedDiscount codedDiscount = CodedDiscount.getDiscountById(discountCode);
         Pattern discountCodePattern = Pattern.compile("^(?i)discount code$");
         Matcher discountCodeMatcher = discountCodePattern.matcher(field);
@@ -178,19 +217,20 @@ public class BossProcessor extends Processor {
             codedDiscount.setDiscountAmount(changeField);
         } else if (remainingMatcher.matches()) {
             codedDiscount.setRepeat(changeField);
+        } else {
+            throw new InvalidCommandException("invalid field");
         }
-
-
     }
 
-    public void processRemoveCodedDiscount(String discountCode) {
-        //TODO : error handling
+    public void processRemoveCodedDiscount(String discountCode) throws NullPointerException {
         CodedDiscount codedDiscount = CodedDiscount.getDiscountById(discountCode);
+        if (codedDiscount == null) {
+            throw new NullPointerException("coded discount with this code doesn't exist");
+        }
         ((Manager) user).removeCodedDiscount(codedDiscount);
     }
 
-    public void manageRequests(String command) {
-        //TODO : error handling
+    public void manageRequests(String command) throws InvalidCommandException {
         if (command.equals("back")) {
             return;
         }
@@ -201,37 +241,56 @@ public class BossProcessor extends Processor {
         Pattern declineRequestPattern = Pattern.compile("decline (.+)");
         Matcher declineRequestMatcher = declineRequestPattern.matcher(command);
         if (detailsMatcher.matches()) {
-            viewRequestDetails(detailsMatcher.group(1));
+            try {
+                viewRequestDetails(detailsMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (acceptRequestMatcher.matches()) {
-            acceptRequest(acceptRequestMatcher.group(1));
+            try {
+                acceptRequest(acceptRequestMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (declineRequestMatcher.matches()) {
-            declineRequest(declineRequestMatcher.group(1));
+            try {
+                declineRequest(declineRequestMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else {
+            throw new InvalidCommandException("invalid command");
         }
     }
 
-    public void viewRequestDetails(String requestId) {
-        //TODO : error handling
+    public void viewRequestDetails(String requestId) throws NullPointerException {
         Request request = Manager.getRequestById(requestId);
+        if (request == null) {
+            throw new NullPointerException("request with this Id doesn't exist");
+        }
         bossViewManager.viewRequestDetails(request);
 
     }
 
-    public void acceptRequest(String requestId) {
-        //TODO : error handling
+    public void acceptRequest(String requestId) throws NullPointerException {
         Request request = Manager.getRequestById(requestId);
+        if (request == null) {
+            throw new NullPointerException("request with this Id doesn't exist");
+        }
         ((Manager) user).acceptRequest(request);
     }
 
-    public void declineRequest(String requestId) {
-        //TODO : error handling
+    public void declineRequest(String requestId) throws NullPointerException {
         Request request = Manager.getRequestById(requestId);
+        if (request == null) {
+            throw new NullPointerException("request with this Id doesn't exist");
+        }
         ((Manager) user).declineRequest(request);
 
     }
 
 
-    public void manageCategories(String command) {
-        //TODO : error handling
+    public void manageCategories(String command) throws InvalidCommandException {
         if (command.equals("back")) {
             return;
         }
@@ -242,38 +301,80 @@ public class BossProcessor extends Processor {
         Pattern removeCategoryPattern = Pattern.compile("remove (.+)");
         Matcher removeCategoryMatcher = removeCategoryPattern.matcher(command);
         if (editCategoryMatcher.matches()) {
-            editCategory(editCategoryMatcher.group(1));
+            try {
+                editCategory(editCategoryMatcher.group(1));
+            } catch (InvalidCommandException | NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (addCategoryMatcher.matches()) {
-            addCategory(addCategoryMatcher.group(1));
+            try {
+                addCategory(addCategoryMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
         } else if (removeCategoryMatcher.matches()) {
-            removeCategory(removeCategoryMatcher.group(1));
+            try {
+                removeCategory(removeCategoryMatcher.group(1));
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else {
+            throw new InvalidCommandException("invalid command");
         }
     }
 
-    public void editCategory(String categoryName) {
-        //TODO : error handling
+    public void editCategory(String categoryName) throws InvalidCommandException {
         Category category = Category.getCategoryByName(categoryName);
+        if (category == null) {
+            throw new NullPointerException("category with this name doesn't exist");
+        }
+        String editTask = bossViewManager.editCategoryTask();
+        if (editTask.equalsIgnoreCase("1")) {
+            String newName = bossViewManager.getCategoryNewName();
+            try {
+                ((Manager) user).editCategoryName(category, newName);
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else if (editTask.equalsIgnoreCase("2")) {
+            String oldFeatureName = bossViewManager.getFeatureNameForChangeOrDelete(category);
+            String newFeatureName = bossViewManager.getFeatureNameForAddOrChange();
+            try {
+                ((Manager) user).editFeatureName(category, oldFeatureName, newFeatureName);
+            } catch (InvalidCommandException | NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else if (editTask.equalsIgnoreCase("3")) {
+            String featureName = bossViewManager.getFeatureNameForChangeOrDelete(category);
+            try {
+                ((Manager) user).deleteFeature(category, featureName);
+            } catch (NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else if (editTask.equalsIgnoreCase("4")) {
+            String featureName = bossViewManager.getFeatureNameForAddOrChange();
+            try {
+                ((Manager) user).addCategoryFeature(category, featureName);
+            } catch (InvalidCommandException | NullPointerException e) {
+                viewManager.showErrorMessage(e.getMessage());
+            }
+        } else {
+            throw new InvalidCommandException("invalid number");
+        }
 
-        ///////////////////////////////////
     }
 
     public void addCategory(String categoryName) {
-        //TODO : error handling
-        ArrayList<String> categoryInfo = new ArrayList<>();
-        ArrayList<String>features = bossViewManager.getCategoryInfo(categoryInfo);
-        if(categoryInfo.size()==2){
-            Category superCategory = Category.getCategoryByName(categoryInfo.get(1));
-            new Category(categoryInfo.get(0), superCategory, features);
-        } else {
-            new Category(categoryInfo.get(0), null, features);
-        }
+        ArrayList<String> features = bossViewManager.getCategoryFeatures();
+        new Category(categoryName, features);
     }
 
-    public void removeCategory(String categoryName) {
-        //TODO : error handling
+    public void removeCategory(String categoryName) throws NullPointerException {
         Category category = Category.getCategoryByName(categoryName);
-        ((Manager)user).removeCategory(category);
-
+        if (category == null) {
+            throw new NullPointerException("category with this name doesn't exist");
+        }
+        ((Manager) user).removeCategory(category);
     }
